+++ CHANGELOG.md
@@ -7,9 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
-- Version number now shown in app title bar.
-- App now starts up maximized on screen.
-
+- Ported from NinjaTerm being an Electron app to a PWA (Progressive Web App).
 ## [3.2.1] - 2023-09-17
 
 - Fixed rendering bug where window resizing would not properly adjust terminal panes.