+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Fixed rendering bug where window resizing would not properly adjust terminal panes.
 - Fixed bug where TX data was not being displayed correctly in the terminals.
+- Removed status message pane from the bottom of the app.
 
 ## [3.2.0] - 2023-09-16