+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+- Nothing here yet
+
+## [4.0.0] - 2023-09-22
+
 - Ported from NinjaTerm being an Electron app to a PWA (Progressive Web App).
 
 ## [3.2.1] - 2023-09-17